fix palette remapper - palette remapper must account for preserved blacks...
some images when converted from EPF to PNG have blacks that arent in the first palette index, and thus arent transparent

when this image is then converted back into an EPF, the black is preserved by changing it to a near-black color

when remapping palettes, we must take this possibility into account. It will show itself as the near-black color being in one palette, but not the other. With this we can guess that it was a preserved black, and we can re-map it back to a black that isnt in the target palette's 0 index.